requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
opencv-python==4.8.1.78
mediapipe==0.10.9
torch==2.1.1
torchvision==0.16.1
numpy==1.24.3
Pillow==10.1.0
scikit-learn==1.3.2
scipy==1.11.4
python-dotenv==1.0.0
